requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
cmdstanpy==1.2.4
contourpy==1.3.0
cycler==0.12.1
Django==5.1
djangorestframework==3.15.2
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
git-filter-repo==2.45.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
holidays==0.56
idna==3.10
importlib_resources==6.4.4
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
lstm==0.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
myapp==0.1.dev0
namex==0.0.8
numpy==1.24.4
nvidia-nccl-cu12==2.22.3
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.0
prophet==1.1.5
protobuf==4.25.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rich==13.8.1
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
sequential==1.0.0
six==1.16.0
sqlparse==0.5.1
stanio==0.5.1
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
views==0.3
Werkzeug==3.0.4
wrapt==1.16.0
xgboost==2.1.1
